requirements.txt
beautifulsoup4==4.6.0
boto3==1.4.8
botocore==1.8.1
certifi==2017.11.5
chardet==3.0.4
docutils==0.14
enum34==1.1.6
fbchat==1.1.1
futures==3.1.1
idna==2.6
jmespath==0.9.3
lxml==4.1.1
olefile==0.44
Pillow==4.3.0
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
urllib3==1.22
